config renamed to cas-4.0.x.properties
some of the new features we have been adding are in fact incompatible with cas-4.0.1, so the name of the config (well in fact the name of this github repo are misleading/confusing).  Mainly the upgrade to pac4j-1.7.x requires cas-4.0.4; previous versions of cas-4.0.x up to cas-4.0.3 were restricted to pac4j-1.4.x.